+++ R/init.R
@@ -12,11 +12,6 @@
 #' @import units
 NULL
 
-# re-export:
-#' @importFrom lwgeom st_transform_proj
-#' @export
-lwgeom::st_transform_proj
-
 setOldClass("stars")
 setOldClass("stars_proxy")